Changed async to future/blocking and changed the error to warn (#353)
Changed the usage of async to Future/blocking. Reason being due to the async keyword, it wasn't able to allocate threads to the process and hence, was timing out very frequently.
The number of concurrently blocking computations can exceed the parallelism level only if each blocking call is wrapped inside a blocking call. Otherwise, there is a risk that the thread pool in the global execution context is starved, and no computation can proceed.
Reference: https://docs.scala-lang.org/overviews/core/futures.html#the-global-execution-context